Check drivers version in khiops (#527)
* Add Utils.h/cpp in Norm

Move GetMajorVersion(const ALString& sFullVersion) and GetMinorVersion(const ALString& sFullVersion)
from Learning library to a new file Utils.h in Norm

* Check the versions of the drivers when loading them

* Drivers messages are warnings
